requirements.txt
Flask==2.3.3
Flask-Cors
Flask-JWT-Extended
Flask-Marshmallow==0.14.0